✨ feat: Add Helm chart for k8s-app
This commit adds the necessary files and configurations to create a Helm chart for the k8s-app. The chart includes templates for deployment, service, ingress, service account, horizontal pod autoscaler (HPA), and test connection. It also includes default values in the values.yaml file for easy customization.

The added files include:
- .helmignore: Specifies patterns to ignore when building packages.
- Chart.lock: Lock file specifying dependencies.
- Chart.yaml: YAML file containing metadata about the chart.
- charts/redis-18.17.0.tgz: Binary file of Redis dependency.
- templates/_helpers.tpl: Helper template with common functions used by other templates.
- templates/deployment.yaml: Deployment configuration template.
- templates/hpa.yaml: HorizontalPodAutoscaler configuration template.
- templates/ingress.yaml: Ingress configuration template.
- templates/NOTES.txt: Template providing instructions on accessing the application URL after deployment.
- templates/service.yaml: Service configuration template.
- templates/serviceaccount.yaml: ServiceAccount configuration template.
- templates/tests/test-connection.yaml: Test connection Pod definition used during installation testing.

Additionally, a values.yaml file is included with default values that can be overridden during installation.

The Redis dependency is enabled by default and has a predefined password.