✨🔧 Add ingress configuration for WordPress installation
This commit adds the necessary configuration for the ingress resource in order to access the WordPress installation. The hostname is set to "nico.k8sw.devninjas.io" and TLS is enabled with annotations for nginx and cert-manager. Additionally, the service type is set to ClusterIP. The default values for the WordPress blog name, password, and username are also provided.

[Reference: Kubernetes Ingress Documentation](https://kubernetes.io/docs/concepts/services-networking/ingress/)