✨📝 Added cluster issuer for Let's Encrypt production
This commit adds a new cluster issuer for Let's Encrypt production. The `cluster-issuer-http.yaml` file now includes the necessary configuration for the `letsencrypt-prod` issuer, which uses the ACME protocol with HTTP01 challenge validation. The ingress class name is set to `nginx`. The email associated with this issuer is nico@devninjas.io.